Fix Extension typing Move heartbeat to view-controller
