+++ packages/magic-sdk/test/spec/iframe-controller/_post.spec.ts
@@ -14,7 +14,7 @@ test('Calls iframe.contentWindow.postMessage with the expected arguments', async
 
   const postMessageStub = jest.fn();
 
-  overlay.checkIframeExistsInDOM = jest.fn().mockResolvedValue({ contentWindow: { postMessage: postMessageStub } });
+  overlay.checkRelayerExistsInDOM = jest.fn().mockResolvedValue({ contentWindow: { postMessage: postMessageStub } });
 
   await (overlay as any)._post({ thisIsData: 'hello world' });
